Add files via upload
This is an hotel cancelation prediction model with 35000 data points it includes all descriptive analysis univariate bivariate and multivariate . After descriptive analysis i further proceeded with statistical analysis first of all checked if the data was normal so that we can could come to conclusion that we should use parametric or non parametric test and we found that the data was not normal and proceeded with mannwhitneyu test and after that we did hypothesis testing of all columns to check if the columns are directly dependent on target or not. If not dependent then we dropped those columns and did not take them in model. Thereafter applied all the algorithm of classification and found out random forest model was the best of all with 83% accuracy.